auto follow Instagram acounts
Open up the Instagram website on your computer.
Log in to your account.
Open up the web browser's developer console by right-clicking on the page and selecting "Inspect" or "Inspect Element".
Find the "Console" tab in the developer console.
Copy and paste the code that you want to use from a trusted source into the console.
Press the "Enter" key on your keyboard to run the code.
Wait for the code to finish running. It should automatically start following accounts on Instagram.
If you want to stop the code from running, close the developer console or refresh the page.
Note: It's important to only use code that you trust from reputable sources. Also, be aware that using automation tools like this on Instagram can violate their terms of service and may result in your account being suspended or banned.